Wider handle for 64 bits arch

+++ include/mm_types.h
@@ -59,7 +59,7 @@
 /**
  * Multimedia Framework handle type.
  */
-typedef int MMHandleType;		/**< Handle type */
+typedef intptr_t MMHandleType;		/**< Handle type */
 
 
 /**
